Fix warnings: Implicit narrowing conversion in compound assignment
- (int count - int pos) here is always an int so amount is also in the
int range if the above test is true.
- We can safely cast and avoid static analysis warnings: "Implicit
narrowing conversion in compound assignment"
- https://github.com/apache/commons-io/security/code-scanning/135
- https://github.com/apache/commons-io/security/code-scanning/88

+++ src/main/java/org/apache/commons/io/input/UnsynchronizedBufferedInputStream.java
@@ -397,10 +397,12 @@ public final class UnsynchronizedBufferedInputStream extends UnsynchronizedFilte
         }
 
         if (count - pos >= amount) {
-            pos += amount;
+            // (int count - int pos) here is always an int so amount is also in the int range if the above test is true.
+            // We can safely cast to int and avoid static analysis warnings.
+            pos += (int) amount;
             return amount;
         }
-        long read = count - pos;
+        int read = count - pos;
         pos = count;
 
         if (markPos != IOUtils.EOF && amount <= markLimit) {
@@ -408,7 +410,9 @@ public final class UnsynchronizedBufferedInputStream extends UnsynchronizedFilte
                 return read;
             }
             if (count - pos >= amount - read) {
-                pos += amount - read;
+                // (int count - int pos) here is always an int so (amount - read) is also in the int range if the above test is true.
+                // We can safely cast to int and avoid static analysis warnings.
+                pos += ((int) amount) - read;
                 return amount;
             }
             // Couldn't get all the bytes, skip what we read
